build: move all runnable Go packages in a single place
