+++ cmd/wardend/cmd/root.go
